Added temperature safety and tests.

+++ TeensyV2/common/core/module.cpp
@@ -86,8 +86,8 @@ void Module::checkPerformanceAndSafety() {
 void Module::getPerformanceStats(char* stats_json, size_t buffer_size) {
   // Start JSON object
   int written = snprintf(stats_json, buffer_size,
-                         "{\"loop_count\":%lu,\"freq\":%.1f,\"modules\":[",
-                         total_loop_count_,
+                         "{\"loop_count\":%u,\"freq\":%.1f,\"modules\":[",
+                         static_cast<unsigned int>(total_loop_count_),
                          total_loop_count_ > 0
                              ? total_loop_count_ * 1000.0f / millis()
                              : 0.0f);
